+++ app/controllers/admin/products_controller.rb
@@ -1,9 +1,4 @@
-class Admin::ProductsController < ApplicationController
-  layout 'admin'
-
-  before_action :authenticate_user!
-  before_action :admin_required
-
+class Admin::ProductsController < AdminController
   def index
     @products = Product.all
   end
