+++ app/controllers/concerns/tokenable.rb
@@ -1,0 +1,13 @@
+module Tokenable
+  extend ActiveSupport::Concern
+
+  included do
+    before_create :generate_token
+  end
+
+  private
+
+  def generate_token
+    self.token = SecureRandom.uuid
+  end
+end
