+++ test/controllers/admin_controller_test.rb
@@ -1,0 +1,7 @@
+require 'test_helper'
+
+class AdminControllerTest < ActionController::TestCase
+  # test "the truth" do
+  #   assert true
+  # end
+end
